# Slim down percentile neighbourhood