@BelligerG, here's a stab at reducing memory of percentile neighbourhood that could be merged into your metoppv#1212 (uses the same idea as for the lapse rate).

Acceptance and unit tests pass, except for one unit test failure (asymmetric kernel). I'm not sure if the original implementation gives the correct result in that case (CC: @fionaRust).