@TomekTrzeciak 2020-04-25
Performance:
Master: 68s walltime (9.5s system), 16GB max RSS
This PR: 34s walltime (0.4s system), 137MB max RSS